Refactor Image Api
