+++ internal/test-utils/src/nonAndroidMain/kotlin/coil3/test/utils/FakeImage.kt
@@ -1,7 +1,7 @@
 package coil3.test.utils
 
-import coil3.CoilPainter
 import coil3.Image
+import org.jetbrains.skia.Canvas
 
 actual class FakeImage actual constructor(
     override val width: Int,
@@ -9,7 +9,5 @@ actual class FakeImage actual constructor(
     override val size: Long,
     override val shareable: Boolean,
 ) : Image {
-    override fun asPainter(): CoilPainter {
-        throw UnsupportedOperationException()
-    }
+    override fun Canvas.onDraw() = Unit
 }
